Refactor Status box
